+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+# 0.12.1 - 2017-11-02
+
 - [server] Add idle event sources
 - [sys] More robust loading of wayland-egl and wayland-cursor (failed on ubuntu 17.04)